ref(tooling): Introduce Makefile and database (#7)
See API-2440

This PR adds a Postgres image to the build. Right now it is not connected to either of the backend applications but it does allow for seeding data. With the help of faker.js, there is no need for JSON any more either.

The instructions are accessible on make or any of the updated READMEs