chore(ts): Add some CRUD routes (#26)
See API-2581

This PR adds the CRUD routes for the TS app along with their tests. It also ensures that both backends work with the frontends by having identical routes/behavior, along with CORS configured. The frontend app has been slightly modified to require an organization with which to view a Kanban board for (e.g. instead of Kanban on localhost:3000, it is on localhost:3000/orgslug/)