feat(ui): Add a landing page, and more info/buttons on cards (#27)
Since as of #26 the app is actually rendered based on the org slug, we should direct users to that page when they visit the home route of the application, rather than showing a blank page. Instead of designing it from scratch, I just lifted most of it from the SetupPage.