feat(webhooks): Explicitly ignore error.created webhooks (#37)
See API-2604

So the motivation behind this PR is to explicitly not take action on error.created webhooks. There isn't a reasonable action the app can take that would take advantage of this webhook, and it would only be available for integrators at a certain plan. I think having a note in the code is sufficient for developers to extend it as they need, while knowing the limitations of that webhook.

If we'd prefer to have an app feature, that works as well, but I couldn't think of something reasonable that wouldn't clog the UI or server logs (e.g. printing a statement, or creating an item in the kanban).