chore(auth): Show necessary hack for current auth workflow (#39)
This PR adds back the signature verification for all incoming requests. Unfortunately, it looks pretty hacky and fragile, but it is what some of our partners have to do if we don't migrate the scheme to a single header, and actual JSON payload.

This PR should serve as an argument in favour of standardizing the headers and payloads on our results.

For more info on how this occurs, and why it's not great:

Sentry is sending a application/json header, with a empty body (b'' in python). In Flask, this will automatically respond with a 400 if you attempt request.json, and in Express, it automatically coerces the empty body into {} since that is valid JSON.

Since we use the app secret and the request body to create the header, this leads to inconsistency with how developers are expected to parse the body, forcing them to always use the raw body since early stage middlewares or helpers (request.json, bodyParser.json()) might alter the payload.

It's difficult to do that in JavaScript since the rest of the API consumes JSON requests, so I manually have to check for req.body === {} and convert it to '' . In Flask, I can't use request.json and have to request.data to access the raw byte string, but this is only because the header isn't telling the truth about the Content-Type.

Additionally, due to a typo in Sentry code, we have two headers which may or may not have the signature to verify against.